Add Qry115A, Qry116AFrm
Edit StopYasrefByTBLBeanWarsaFrm, TblAttachEditFrm, TblWarasaAmanat and other

+++ RetirementCenter/Forms/Data/TblWarasaAmanatFrm.cs
@@ -124,6 +124,11 @@ namespace RetirementCenter
                     Program.Logger.LogThis("لا يمكن حذف عنصر تم مراجعتة", Text, FXFW.Logger.OpType.warning, null, null, this);
                     return;
                 }
+                if (!row.IsdateinNull() && row.datein != SQLProvider.ServerDateTime() && Program.UserInfo.IsAdmin == false)
+                {
+                    msgDlg.Show("الحذف متاح في يوم الادخال فقط", msgDlg.msgButtons.Close);
+                    return;
+                }
                 tblWarasaAmanatTableAdapter.Delete(row.PersonId, row.DofatSarfAId, row.amanattypeid);
                 gridViewData.DeleteRow(GV.FocusedRowHandle);
